+++ banyand/stream/query.go
@@ -24,21 +24,24 @@ import (
 	"fmt"
 	"sort"
 
-	"go.uber.org/multierr"
-
 	"github.com/apache/skywalking-banyandb/api/common"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
 	"github.com/apache/skywalking-banyandb/banyand/internal/storage"
 	"github.com/apache/skywalking-banyandb/pkg/convert"
+	"github.com/apache/skywalking-banyandb/pkg/index"
 	itersort "github.com/apache/skywalking-banyandb/pkg/iter/sort"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/partition"
 	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
+	"github.com/apache/skywalking-banyandb/pkg/query/logical"
+	"github.com/apache/skywalking-banyandb/pkg/timestamp"
 )
 
+type seriesTimestampMap map[common.SeriesID][]int64
+
 type queryOptions struct {
-	elementRefMap map[common.SeriesID][]int64
+	filteredRefMap seriesTimestampMap
 	pbv1.StreamQueryOptions
 	minTimestamp int64
 	maxTimestamp int64
@@ -140,16 +143,17 @@ func strArrTagValue(values []string) *modelv1.TagValue {
 }
 
 type queryResult struct {
-	entityMap    map[string]int
-	sidToIndex   map[common.SeriesID]int
-	tagNameIndex map[string]partition.TagLocator
-	schema       *databasev1.Stream
-	data         []*blockCursor
-	snapshots    []*snapshot
-	seriesList   pbv1.SeriesList
-	loaded       bool
-	orderByTS    bool
-	ascTS        bool
+	entityMap      map[string]int
+	sidToIndex     map[common.SeriesID]int
+	tagNameIndex   map[string]partition.TagLocator
+	schema         *databasev1.Stream
+	elementRefList []*elementRef
+	data           []*blockCursor
+	snapshots      []*snapshot
+	seriesList     pbv1.SeriesList
+	loaded         bool
+	orderByTS      bool
+	asc            bool
 }
 
 func (qr *queryResult) Pull() *pbv1.StreamResult {
@@ -231,8 +235,12 @@ func (qr *queryResult) Pull() *pbv1.StreamResult {
 		qr.loaded = true
 		heap.Init(qr)
 	}
+
 	if len(qr.data) == 0 {
 		return nil
+	}
+	if !qr.orderByTS {
+		return qr.mergeByTagValue()
 	}
 	if len(qr.data) == 1 {
 		r := &pbv1.StreamResult{}
@@ -241,7 +249,7 @@ func (qr *queryResult) Pull() *pbv1.StreamResult {
 		qr.data = qr.data[:0]
 		return r
 	}
-	return qr.merge()
+	return qr.mergeByTimestamp()
 }
 
 func (qr *queryResult) Release() {
@@ -261,17 +269,13 @@ func (qr queryResult) Len() int {
 }
 
 func (qr queryResult) Less(i, j int) bool {
-	leftTS := qr.data[i].timestamps[qr.data[i].idx]
-	rightTS := qr.data[j].timestamps[qr.data[j].idx]
-	if qr.orderByTS {
-		if qr.ascTS {
-			return leftTS < rightTS
-		}
-		return leftTS > rightTS
+	leftIdx, rightIdx := qr.data[i].idx, qr.data[j].idx
+	leftTS := qr.data[i].timestamps[leftIdx]
+	rightTS := qr.data[j].timestamps[rightIdx]
+	if qr.asc {
+		return leftTS < rightTS
 	}
-	leftSIDIndex := qr.sidToIndex[qr.data[i].bm.seriesID]
-	rightSIDIndex := qr.sidToIndex[qr.data[j].bm.seriesID]
-	return leftSIDIndex < rightSIDIndex
+	return leftTS > rightTS
 }
 
 func (qr queryResult) Swap(i, j int) {
@@ -291,10 +295,59 @@ func (qr *queryResult) Pop() interface{} {
 }
 
 func (qr *queryResult) orderByTimestampDesc() bool {
-	return qr.orderByTS && !qr.ascTS
+	return qr.orderByTS && !qr.asc
 }
 
-func (qr *queryResult) merge() *pbv1.StreamResult {
+func (qr *queryResult) mergeByTagValue() *pbv1.StreamResult {
+	tmp := &pbv1.StreamResult{}
+	prevIdx := 0
+	elementRefToIdx := make(map[elementRef]int)
+	for _, data := range qr.data {
+		data.copyAllTo(tmp, qr.orderByTimestampDesc())
+		var idx int
+		for idx = prevIdx; idx < len(tmp.Timestamps); idx++ {
+			sid, ts := tmp.SIDs[idx], tmp.Timestamps[idx]
+			er := elementRef{
+				seriesID:  sid,
+				timestamp: ts,
+			}
+			elementRefToIdx[er] = idx
+		}
+		prevIdx = idx
+	}
+
+	r := &pbv1.StreamResult{
+		TagFamilies: []pbv1.TagFamily{},
+	}
+	for _, tagFamily := range tmp.TagFamilies {
+		tf := pbv1.TagFamily{
+			Name: tagFamily.Name,
+			Tags: []pbv1.Tag{},
+		}
+		for _, tag := range tagFamily.Tags {
+			t := pbv1.Tag{
+				Name:   tag.Name,
+				Values: []*modelv1.TagValue{},
+			}
+			tf.Tags = append(tf.Tags, t)
+		}
+		r.TagFamilies = append(r.TagFamilies, tf)
+	}
+	for _, er := range qr.elementRefList {
+		idx := elementRefToIdx[*er]
+		r.Timestamps = append(r.Timestamps, tmp.Timestamps[idx])
+		r.ElementIDs = append(r.ElementIDs, tmp.ElementIDs[idx])
+		for i := 0; i < len(r.TagFamilies); i++ {
+			for j := 0; j < len(r.TagFamilies[i].Tags); j++ {
+				r.TagFamilies[i].Tags[j].Values = append(r.TagFamilies[i].Tags[j].Values, tmp.TagFamilies[i].Tags[j].Values[idx])
+			}
+		}
+	}
+	qr.data = qr.data[:0]
+	return r
+}
+
+func (qr *queryResult) mergeByTimestamp() *pbv1.StreamResult {
 	step := 1
 	if qr.orderByTimestampDesc() {
 		step = -1
@@ -330,218 +383,7 @@ func (qr *queryResult) merge() *pbv1.StreamResult {
 	return result
 }
 
-func (s *stream) genIndex(tagProj []pbv1.TagProjection, seriesList pbv1.SeriesList) (map[string]int, map[string]*databasev1.TagSpec,
-	map[string]partition.TagLocator, map[common.SeriesID]int,
-) {
-	entityMap := make(map[string]int)
-	for idx, entity := range s.schema.GetEntity().GetTagNames() {
-		entityMap[entity] = idx + 1
-	}
-	tagSpecIndex := make(map[string]*databasev1.TagSpec)
-	for _, tagFamilySpec := range s.schema.GetTagFamilies() {
-		for _, tagSpec := range tagFamilySpec.Tags {
-			tagSpecIndex[tagSpec.GetName()] = tagSpec
-		}
-	}
-	tagProjIndex := make(map[string]partition.TagLocator)
-	for i, tagFamilyProj := range tagProj {
-		for j, tagProj := range tagFamilyProj.Names {
-			if entityMap[tagProj] == 0 {
-				continue
-			}
-			tagProjIndex[tagProj] = partition.TagLocator{
-				FamilyOffset: i,
-				TagOffset:    j,
-			}
-		}
-	}
-	sidToIndex := make(map[common.SeriesID]int)
-	for idx, series := range seriesList {
-		sidToIndex[series.ID] = idx
-	}
-
-	return entityMap, tagSpecIndex, tagProjIndex, sidToIndex
-}
-
-func (s *stream) Query(ctx context.Context, sqo pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error) {
-	if sqo.TimeRange == nil || len(sqo.Entities) < 1 {
-		return nil, errors.New("invalid query options: timeRange and series are required")
-	}
-	if len(sqo.TagProjection) == 0 {
-		return nil, errors.New("invalid query options: tagProjection is required")
-	}
-	db := s.databaseSupplier.SupplyTSDB()
-	var result queryResult
-	if db == nil {
-		return &result, nil
-	}
-	tsdb := db.(storage.TSDB[*tsTable, option])
-	tabWrappers := tsdb.SelectTSTables(*sqo.TimeRange)
-	defer func() {
-		for i := range tabWrappers {
-			tabWrappers[i].DecRef()
-		}
-	}()
-	series := make([]*pbv1.Series, len(sqo.Entities))
-	for i := range sqo.Entities {
-		series[i] = &pbv1.Series{
-			Subject:      sqo.Name,
-			EntityValues: sqo.Entities[i],
-		}
-	}
-	sl, err := tsdb.Lookup(ctx, series)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(sl) < 1 {
-		return &result, nil
-	}
-	var sids []common.SeriesID
-	for i := range sl {
-		sids = append(sids, sl[i].ID)
-	}
-	var parts []*part
-	qo := queryOptions{
-		StreamQueryOptions: sqo,
-		minTimestamp:       sqo.TimeRange.Start.UnixNano(),
-		maxTimestamp:       sqo.TimeRange.End.UnixNano(),
-	}
-	var n int
-	for i := range tabWrappers {
-		s := tabWrappers[i].Table().currentSnapshot()
-		if s == nil {
-			continue
-		}
-		parts, n = s.getParts(parts, qo.minTimestamp, qo.maxTimestamp)
-		if n < 1 {
-			s.decRef()
-			continue
-		}
-		result.snapshots = append(result.snapshots, s)
-	}
-	bma := generateBlockMetadataArray()
-	defer releaseBlockMetadataArray(bma)
-	// TODO: cache tstIter
-	var ti tstIter
-	defer ti.reset()
-	originalSids := make([]common.SeriesID, len(sids))
-	copy(originalSids, sids)
-	sort.Slice(sids, func(i, j int) bool { return sids[i] < sids[j] })
-	ti.init(bma, parts, sids, qo.minTimestamp, qo.maxTimestamp)
-	if ti.Error() != nil {
-		return nil, fmt.Errorf("cannot init tstIter: %w", ti.Error())
-	}
-	for ti.nextBlock() {
-		bc := generateBlockCursor()
-		p := ti.piHeap[0]
-		bc.init(p.p, p.curBlock, qo)
-		result.data = append(result.data, bc)
-	}
-	if ti.Error() != nil {
-		return nil, fmt.Errorf("cannot iterate tstIter: %w", ti.Error())
-	}
-
-	entityMap, _, _, sidToIndex := s.genIndex(sqo.TagProjection, sl)
-	result.entityMap = entityMap
-	result.sidToIndex = sidToIndex
-	result.tagNameIndex = make(map[string]partition.TagLocator)
-	result.schema = s.schema
-	result.seriesList = sl
-	for i, si := range originalSids {
-		result.sidToIndex[si] = i
-	}
-	for i, tagFamilySpec := range s.schema.GetTagFamilies() {
-		for j, tagSpec := range tagFamilySpec.GetTags() {
-			result.tagNameIndex[tagSpec.GetName()] = partition.TagLocator{
-				FamilyOffset: i,
-				TagOffset:    j,
-			}
-		}
-	}
-	result.orderByTS = true
-	if sqo.Order == nil {
-		result.ascTS = true
-		return &result, nil
-	}
-	if sqo.Order.Sort == modelv1.Sort_SORT_ASC || sqo.Order.Sort == modelv1.Sort_SORT_UNSPECIFIED {
-		result.ascTS = true
-	}
-	return &result, nil
-}
-
-func (s *stream) Sort(ctx context.Context, sqo pbv1.StreamQueryOptions) (ssr pbv1.StreamSortResult, err error) {
-	if sqo.TimeRange == nil || len(sqo.Entities) < 1 {
-		return nil, errors.New("invalid query options: timeRange and series are required")
-	}
-	if len(sqo.TagProjection) == 0 {
-		return nil, errors.New("invalid query options: tagProjection is required")
-	}
-	db := s.databaseSupplier.SupplyTSDB()
-	if db == nil {
-		return ssr, nil
-	}
-	tsdb := db.(storage.TSDB[*tsTable, option])
-	tabWrappers := tsdb.SelectTSTables(*sqo.TimeRange)
-	defer func() {
-		for i := range tabWrappers {
-			tabWrappers[i].DecRef()
-		}
-	}()
-
-	series := make([]*pbv1.Series, len(sqo.Entities))
-	for i := range sqo.Entities {
-		series[i] = &pbv1.Series{
-			Subject:      sqo.Name,
-			EntityValues: sqo.Entities[i],
-		}
-	}
-	seriesList, err := tsdb.Lookup(ctx, series)
-	if err != nil {
-		return nil, err
-	}
-	if len(seriesList) == 0 {
-		return ssr, nil
-	}
-
-	iters, err := s.buildSeriesByIndex(tabWrappers, seriesList, sqo)
-	if err != nil {
-		return nil, err
-	}
-	if len(iters) == 0 {
-		return ssr, nil
-	}
-
-	it := newItemIter(iters, sqo.Order.Sort)
-	defer func() {
-		err = multierr.Append(err, it.Close())
-	}()
-
-	ces := newColumnElements()
-	for it.Next() {
-		nextItem := it.Val()
-		e := nextItem.element
-		ces.BuildFromElement(e, sqo.TagProjection)
-		if len(ces.timestamp) >= sqo.MaxElementSize {
-			break
-		}
-	}
-	return ces, err
-}
-
-// newItemIter returns a ItemIterator which mergers several tsdb.Iterator by input sorting order.
-func newItemIter(iters []*searcherIterator, s modelv1.Sort) itersort.Iterator[item] {
-	var ii []itersort.Iterator[item]
-	for _, iter := range iters {
-		ii = append(ii, iter)
-	}
-	if s == modelv1.Sort_SORT_DESC {
-		return itersort.NewItemIter[item](ii, true)
-	}
-	return itersort.NewItemIter[item](ii, false)
-}
-
-func (s *stream) Filter(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr pbv1.StreamQueryResult, err error) {
+func (s *stream) Query(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr pbv1.StreamQueryResult, err error) {
 	if sqo.TimeRange == nil || len(sqo.Entities) < 1 {
 		return nil, errors.New("invalid query options: timeRange and series are required")
 	}
@@ -576,37 +418,24 @@ func (s *stream) Filter(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr p
 		return sqr, nil
 	}
 
-	var elementRefList []elementRef
-	for _, tw := range tabWrappers {
-		index := tw.Table().Index()
-		erl, err := index.Search(ctx, seriesList, sqo.Filter, sqo.TimeRange, sqo.Order)
-		if err != nil {
-			return nil, err
-		}
-		elementRefList = append(elementRefList, erl...)
-		if len(elementRefList) > sqo.MaxElementSize {
-			elementRefList = elementRefList[:sqo.MaxElementSize]
-			break
-		}
+	filteredRefMap, err := indexSearch(ctx, sqo, tabWrappers, seriesList)
+	if err != nil {
+		return nil, err
 	}
-	var elementRefMap map[common.SeriesID][]int64
-	if len(elementRefList) != 0 {
-		elementRefMap = make(map[common.SeriesID][]int64)
-		for _, ref := range elementRefList {
-			if _, ok := elementRefMap[ref.seriesID]; !ok {
-				elementRefMap[ref.seriesID] = []int64{ref.timestamp}
-			} else {
-				elementRefMap[ref.seriesID] = append(elementRefMap[ref.seriesID], ref.timestamp)
-			}
-		}
+	elementRefList, sortedRefMap, err := indexSort(s, sqo, tabWrappers, seriesList, filteredRefMap)
+	if err != nil {
+		return nil, err
 	}
+	if sortedRefMap != nil {
+		filteredRefMap = sortedRefMap
+	}
+
 	qo := queryOptions{
 		StreamQueryOptions: sqo,
 		minTimestamp:       sqo.TimeRange.Start.UnixNano(),
 		maxTimestamp:       sqo.TimeRange.End.UnixNano(),
-		elementRefMap:      elementRefMap,
+		filteredRefMap:     filteredRefMap,
 	}
-
 	var parts []*part
 	var n int
 	for i := range tabWrappers {
@@ -629,7 +458,7 @@ func (s *stream) Filter(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr p
 	var sids []common.SeriesID
 	for i := 0; i < len(seriesList); i++ {
 		sid := seriesList[i].ID
-		if _, ok := elementRefMap[sid]; !ok {
+		if filteredRefMap != nil && filteredRefMap[sid] == nil {
 			seriesList = append(seriesList[:i], seriesList[i+1:]...)
 			i--
 			continue
@@ -653,7 +482,7 @@ func (s *stream) Filter(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr p
 		return nil, fmt.Errorf("cannot iterate tstIter: %w", ti.Error())
 	}
 
-	entityMap, _, _, sidToIndex := s.genIndex(sqo.TagProjection, seriesList)
+	entityMap, sidToIndex := s.genIndex(seriesList)
 	result.entityMap = entityMap
 	result.sidToIndex = sidToIndex
 	result.tagNameIndex = make(map[string]partition.TagLocator)
@@ -670,13 +499,122 @@ func (s *stream) Filter(ctx context.Context, sqo pbv1.StreamQueryOptions) (sqr p
 			}
 		}
 	}
-	result.orderByTS = true
 	if sqo.Order == nil {
-		result.ascTS = true
+		result.orderByTS = true
+		result.asc = true
 		return &result, nil
 	}
+	if sqo.Order.Index == nil {
+		result.orderByTS = true
+	} else {
+		result.elementRefList = elementRefList
+	}
 	if sqo.Order.Sort == modelv1.Sort_SORT_ASC || sqo.Order.Sort == modelv1.Sort_SORT_UNSPECIFIED {
-		result.ascTS = true
+		result.asc = true
 	}
 	return &result, nil
+}
+
+func indexSearch(ctx context.Context, sqo pbv1.StreamQueryOptions,
+	tabWrappers []storage.TSTableWrapper[*tsTable], seriesList pbv1.SeriesList,
+) (map[common.SeriesID][]int64, error) {
+	if sqo.Filter == nil || sqo.Filter == logical.ENode {
+		return nil, nil
+	}
+	var filteredRefList []elementRef
+	for _, tw := range tabWrappers {
+		index := tw.Table().Index()
+		erl, err := index.Search(ctx, seriesList, sqo.Filter, sqo.TimeRange, sqo.Order)
+		if err != nil {
+			return nil, err
+		}
+		filteredRefList = append(filteredRefList, erl...)
+		if (sqo.Order == nil || sqo.Order.Index == nil) && len(filteredRefList) > sqo.MaxElementSize {
+			filteredRefList = filteredRefList[:sqo.MaxElementSize]
+			break
+		}
+	}
+	filteredRefMap := make(map[common.SeriesID][]int64)
+	if len(filteredRefList) != 0 {
+		for _, ref := range filteredRefList {
+			if _, ok := filteredRefMap[ref.seriesID]; !ok {
+				filteredRefMap[ref.seriesID] = []int64{ref.timestamp}
+			} else {
+				filteredRefMap[ref.seriesID] = append(filteredRefMap[ref.seriesID], ref.timestamp)
+			}
+		}
+	}
+	return filteredRefMap, nil
+}
+
+func indexSort(s *stream, sqo pbv1.StreamQueryOptions, tabWrappers []storage.TSTableWrapper[*tsTable],
+	seriesList pbv1.SeriesList, filteredRefMap map[common.SeriesID][]int64,
+) ([]*elementRef, map[common.SeriesID][]int64, error) {
+	if sqo.Order == nil || sqo.Order.Index == nil {
+		return nil, nil, nil
+	}
+	iters, err := s.buildItersByIndex(tabWrappers, seriesList, sqo)
+	if err != nil {
+		return nil, nil, err
+	}
+	desc := sqo.Order != nil && sqo.Order.Index == nil && sqo.Order.Sort == modelv1.Sort_SORT_DESC
+	itemIter := itersort.NewItemIter[*index.ItemRef](iters, desc)
+
+	var elementRefList []*elementRef
+	sortedRefMap := make(map[common.SeriesID][]int64)
+	for {
+		if !itemIter.Next() {
+			break
+		}
+		val := itemIter.Val()
+		ts, sid := val.DocID, val.SeriesID
+		if filteredRefMap != nil && (filteredRefMap[sid] == nil || timestamp.Find(filteredRefMap[sid], int64(ts)) == -1) {
+			continue
+		}
+		if _, ok := sortedRefMap[sid]; !ok {
+			sortedRefMap[sid] = []int64{int64(ts)}
+		} else {
+			sortedRefMap[sid] = append(sortedRefMap[sid], int64(ts))
+		}
+		elementRefList = append(elementRefList, &elementRef{timestamp: int64(ts), seriesID: sid})
+		if len(elementRefList) >= sqo.MaxElementSize {
+			break
+		}
+	}
+	return elementRefList, sortedRefMap, nil
+}
+
+func (s *stream) buildItersByIndex(tableWrappers []storage.TSTableWrapper[*tsTable],
+	seriesList pbv1.SeriesList, sqo pbv1.StreamQueryOptions,
+) (iters []itersort.Iterator[*index.ItemRef], err error) {
+	indexRuleForSorting := sqo.Order.Index
+	if len(indexRuleForSorting.Tags) != 1 {
+		return nil, fmt.Errorf("only support one tag for sorting, but got %d", len(indexRuleForSorting.Tags))
+	}
+	sids := seriesList.IDs()
+	for _, tw := range tableWrappers {
+		var iter index.FieldIterator[*index.ItemRef]
+		fieldKey := index.FieldKey{
+			IndexRuleID: indexRuleForSorting.GetMetadata().GetId(),
+			Analyzer:    indexRuleForSorting.GetAnalyzer(),
+		}
+		iter, err = tw.Table().Index().Sort(sids, fieldKey, sqo.Order.Sort, sqo.MaxElementSize)
+		if err != nil {
+			return nil, err
+		}
+		iters = append(iters, iter)
+	}
+	return
+}
+
+func (s *stream) genIndex(seriesList pbv1.SeriesList) (map[string]int, map[common.SeriesID]int) {
+	entityMap := make(map[string]int)
+	for idx, entity := range s.schema.GetEntity().GetTagNames() {
+		entityMap[entity] = idx + 1
+	}
+	sidToIndex := make(map[common.SeriesID]int)
+	for idx, series := range seriesList {
+		sidToIndex[series.ID] = idx
+	}
+	return entityMap, sidToIndex
 }