+++ pkg/index/inverted/inverted.go
@@ -160,7 +160,7 @@ func (s *store) Write(fields []index.Field, docID uint64) error {
 	return nil
 }
 
-func (s *store) Iterator(fieldKey index.FieldKey, termRange index.RangeOpts, order modelv1.Sort, preLoadSize int) (iter index.FieldIterator, err error) {
+func (s *store) Iterator(fieldKey index.FieldKey, termRange index.RangeOpts, order modelv1.Sort, preLoadSize int) (iter index.FieldIterator[*index.ItemRef], err error) {
 	if termRange.Lower != nil &&
 		termRange.Upper != nil &&
 		bytes.Compare(termRange.Lower, termRange.Upper) > 0 {
@@ -241,8 +241,7 @@ func (s *store) MatchTerms(field index.Field) (list posting.List, err error) {
 	}()
 	list = roaring.NewPostingList()
 	for iter.Next() {
-		docID, _, _ := iter.Val()
-		list.Insert(docID)
+		list.Insert(iter.Val().DocID)
 	}
 	return list, err
 }
@@ -275,8 +274,7 @@ func (s *store) Match(fieldKey index.FieldKey, matches []string) (posting.List, 
 	}()
 	list := roaring.NewPostingList()
 	for iter.Next() {
-		docID, _, _ := iter.Val()
-		list.Insert(docID)
+		list.Insert(iter.Val().DocID)
 	}
 	return list, err
 }
@@ -288,8 +286,7 @@ func (s *store) Range(fieldKey index.FieldKey, opts index.RangeOpts) (list posti
 	}
 	list = roaring.NewPostingList()
 	for iter.Next() {
-		docID, _, _ := iter.Val()
-		list.Insert(docID)
+		list.Insert(iter.Val().DocID)
 	}
 	err = multierr.Append(err, iter.Close())
 	return
@@ -456,8 +453,12 @@ func (bmi *blugeMatchIterator) Next() bool {
 	return bmi.err == nil
 }
 
-func (bmi *blugeMatchIterator) Val() (uint64, common.SeriesID, []byte) {
-	return bmi.docID, bmi.seriesID, bmi.term
+func (bmi *blugeMatchIterator) Val() *index.ItemRef {
+	return &index.ItemRef{
+		SeriesID: bmi.seriesID,
+		DocID:    bmi.docID,
+		Term:     bmi.term,
+	}
 }
 
 func (bmi *blugeMatchIterator) Close() error {